pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-data-lab"
version = "0.1.0"
description = "Governed local data pipeline for financial receipts"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Financial Data Lab"}]

[project.scripts]
fdl = "financial_data_lab.cli:main"

[project.optional-dependencies]
test = ["pytest"]
ocr = ["pytesseract", "Pillow"]
ocr_pdf = ["pymupdf"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
